+++ components/DraggablePanel/index.js
@@ -1,0 +1,7 @@
+import DraggablePanel from './draggablePanel';
+
+DraggablePanel.install = function (Vue) {
+    Vue.component(DraggablePanel.name, DraggablePanel);
+};
+
+export default DraggablePanel;